Creating common dir and correcting P4 string functions
