+++ shramko.maksim/P4/countConsecutiveIdenticalSymbols.hpp
@@ -1,0 +1,11 @@
+#ifndef COUNTCONSECUTIVEIDENTICALSYMBOLS_HPP
+#define COUNTCONSECUTIVEIDENTICALSYMBOLS_HPP
+
+#include <cstddef>
+
+namespace shramko
+{
+  size_t countConsecutiveIdenticalSymbols(const char* str);
+}
+
+#endif
