+++ shramko.maksim/common/inputString.hpp
@@ -1,0 +1,13 @@
+#ifndef INPUTSTRING_HPP
+#define INPUTSTRING_HPP
+
+#include <cstddef>
+#include <istream>
+
+namespace shramko
+{
+  char* createBiggerMemory(size_t memory, const char* str);
+  char* inputStr(std::istream& in);
+}
+
+#endif
